implement get_build_by_name_env
Pass it with job name like `DEPLOY-act2-certificate` and the filter
example ['ENV': 'int'] it will return a groovy map of parameters used to build that job
The build needs to be the last success and matching the filter map.

Use case:

The information in the build param can be used to trigger a new job.

The information in the parameter list will have the UPSTREAM_JOB_NAME and
UPSTREAM_JOB_NUMBER which is usefull to trigger another deploy job on defferent
environment for example `tst`.